rename http grpc access logger files
