+++ source/extensions/extensions_build_config.bzl
@@ -5,7 +5,7 @@ EXTENSIONS = {
     #
 
     "envoy.access_loggers.file":                        "//source/extensions/access_loggers/file:config",
-    "envoy.access_loggers.http_grpc":                   "//source/extensions/access_loggers/http_grpc:config",
+    "envoy.access_loggers.http_grpc":                   "//source/extensions/access_loggers/grpc:http_config",
 
     #
     # Clusters
@@ -144,7 +144,7 @@ WINDOWS_EXTENSIONS = {
     #
 
     "envoy.access_loggers.file":                        "//source/extensions/access_loggers/file:config",
-    #"envoy.access_loggers.http_grpc":                   "//source/extensions/access_loggers/http_grpc:config",
+    #"envoy.access_loggers.http_grpc":                   "//source/extensions/access_loggers/grpc:http_config",
 
     #
     # gRPC Credentials Plugins
